requirements.txt
cloudscraper
tabulate
bs4
image-to-ascii-pyaoponto
